+++ examples/index.html
@@ -10,7 +10,7 @@
 
 </div>
 <script type="module">
-	import syntaxHighlight from '../src/syntax-highlight.js';
+	import syntaxHighlight from '../src/syntaxhighlight.js';
 
 	const showExamples = async () => {
 		const example = document.getElementById('examples');
