+++ examples/webcomponent-example.html
@@ -75,7 +75,7 @@ C++
 	}
 </syntax-highlight>
 Using an imported source file and highlight that
-<syntax-highlight height="300px" language="JavaScript" src="../src/syntax-highlight.js">
+<syntax-highlight height="300px" language="JavaScript" src="../src/syntaxhighlight.js">
 </syntax-highlight>
 </body>
 </html>
