+++ src/syntaxhighlight.js
