+++ src/themes/theme-base/README.md
@@ -1,6 +1,6 @@
 # theme-base
 
-Base CSS theme for all [SyntaxHighlight](https://github.com/virtualcodewarrior/syntax-highlight) themes. If you want to create a new theme, please use [theme-django](https://github.com/virtualcodewarrior/syntax-highlight/themes/theme-django) as a starting point.
+Base CSS theme for all [SyntaxHighlight](https://github.com/virtualcodewarrior/syntax-highlight) themes. If you want to create a new theme, please use [theme-django](https://github.com/virtualcodewarrior/syntaxhighlight/themes/theme-django) as a starting point.
 
 ## License
