+++ src/webcomponent/syntax-highlight-worker.js
@@ -1,7 +1,7 @@
 // eslint-disable-next-line no-undef
 let HighlighterModule;
 onmessage = async(event) => {
-	HighlighterModule = HighlighterModule ?? await import('../syntax-highlight.js');
+	HighlighterModule = HighlighterModule ?? await import('../syntaxhighlight.js');
 	// eslint-disable-next-line no-undef
 	const result = await HighlighterModule.default.highlightLanguage(event.data.code, event.data.language, event.data.options);
 	postMessage({ formatted: result, id: event.data.id });
